dont assume register_callsites only called once
